use same docker-compose setup for dev and test
